Moving Literal<T> class to its own file
